Added Ammo Capacity + ADS Movement 🔫
⚠️ NOTE ⚠️

Don't mind the error in the logs saying:

"MissingReferenceException: The object of type 'CheckpointController' has been destroyed but you are still trying to access it.
Your script should either check if it is null or you should not destroy the object.

CheckpointController.Start () (at Assets/Scripts/CheckpointController.cs:29)"

It's not a bug, since the player keeps spawning randomly, the code is trying to find the player at the initial starting position. TL;DR Don't worry, if it works... IT WORKS!